Validação dos dados do perfil de usuário.

+++ app/Models/User.php
@@ -60,7 +60,11 @@ class User extends Authenticatable implements JWTSubject
     public function updateWithProfile(array $data) : User
     {
         try {
-            UserProfile::uploadPhoto($data['photo']);
+            
+            if(isset($data['photo'])){
+                UserProfile::uploadPhoto($data['photo']);
+            }
+            
             \DB::beginTransaction();
             $this->fill($data);
             $this->save();
@@ -68,7 +72,9 @@ class User extends Authenticatable implements JWTSubject
             \DB::commit();
         }catch (\Exception $e){
             //excluir a photo - roll back
-            UserProfile::deleteFile($data['photo']);
+            if(isset($data['photo'])){
+                UserProfile::deleteFile($data['photo']);
+            }
             \DB::rollBack();
             throw $e;
         }
